# openapi.yaml
openapi: 3.0.1
info:
  title: Customer Service API
  description: API for managing customers in the ms-customers microservice.
  version: 0.0.1-SNAPSHOT
servers:
  - url: http://localhost:8080
    description: Local development server

paths:
  /api/customers/create:
    post:
      summary: Create a new customer
      description: Creates a new customer record in the database.
      operationId: createCustomer
      requestBody:
        description: Customer object that needs to be added to the store
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
      responses:
        '200':
          description: Customer created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '503':
          description: Service Unavailable (due to Circuit Breaker or other issues)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/customers/getCustomerTypeById/{id}:
    get:
      summary: Get customer type by ID
      description: Retrieves the type of a specific customer by their ID.
      operationId: getCustomerTypeById
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the customer
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerType'
        '404':
          description: Customer not found
        '503':
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/customers/getById/{id}:
    get:
      summary: Get customer by ID
      description: Retrieves a single customer by their ID.
      operationId: getCustomerById
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the customer to retrieve
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '404':
          description: Customer not found
        '503':
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/customers/all:
    get:
      summary: Get all customers
      description: Retrieves a list of all customers.
      operationId: getAllCustomers
      responses:
        '200':
          description: A list of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customer'
        '503':
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/customers/update:
    put:
      summary: Update an existing customer
      description: Updates the details of an existing customer.
      operationId: updateCustomer
      requestBody:
        description: Customer object that needs to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
      responses:
        '200':
          description: Customer updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '404':
          description: Customer not found
        '503':
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/customers/delete/{id}:
    delete:
      summary: Delete a customer
      description: Deletes a customer by their ID.
      operationId: deleteCustomer
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the customer to delete
          schema:
            type: string
      responses:
        '204':
          description: Customer deleted successfully
        '404':
          description: Customer not found
        '503':
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    Customer:
      type: object
      properties:
        id:
          type: string
          readOnly: true
        customerType:
          $ref: '#/components/schemas/CustomerType'
        documentType:
          type: string
          enum: [DNI, RUC, PASSPORT] # Example values
        documentNumber:
          type: string
        name:
          type: string
        companyName:
          type: string
        email:
          type: string
          format: email
        status:
          type: string
          enum: [ACTIVE, INACTIVE] # Example values
    CustomerType:
      type: string
      enum: [PERSONAL, BUSINESS] # Example values
    Error:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string